feat: Implement AI-powered energy insights and Supabase Edge Functions
Integrates Gemini AI via Supabase Edge Functions to provide personalized energy consumption forecasts and savings recommendations.

- **Supabase Infrastructure:**
    - Initializes Supabase project configuration with API, database, and auth settings.
    - Implements the `get-ai-insights` Edge Function using Deno, which fetches meter logs and queries Gemini AI for forecasts and usage analysis.
    - Installs and configures the Supabase Functions plugin in the mobile client.
- **Insights Screen & ViewModel:**
    - Refactors `InsightsScreen` to support loading states and dynamic data binding.
    - Implements `InsightsScreenViewModel` to calculate local statistics (average spending, daily usage, peak hours, and weekend vs. weekday trends).
    - Adds `loadInsights` to trigger both local calculations and remote AI analysis for a specific meter.
    - Introduces `AIInsightResponse` data model for parsing structured JSON from the Edge Function.
- **HomeViewModel:**
    - Refines consumption simulation math by ensuring explicit double division for `hoursPassed`.
- **Tooling:**
    - Configures VS Code settings and extensions for Deno and TypeScript development within the Supabase environment.

+++ composeApp/src/commonMain/kotlin/com/example/powertracker/viewmodel/homeViewModel.kt
@@ -125,7 +125,7 @@ class HomeViewModel : ViewModel() {
                 val lastLogTime = lastLog?.loggedAt?.let { Instant.parse(it) } ?: (now - 1.hours)
 
                 val duration = now - lastLogTime
-                val hoursPassed = duration.inWholeMilliseconds / 3600000.0
+                val hoursPassed = duration.inWholeMilliseconds.toDouble() / 3600000.0
                 
                 val cappedHours = if (lastLog == null) 1.0 else hoursPassed
